arm64/vmm: Enable load/store of capabilities in some EL2 mappings
We currently map three types of objects into EL2: per-host CPU stacks,
the per-VM hyp structure, and per-vCPU hypctx structures.  The
hypervisor certainly needs to be able to store and load capabilities
from the stack, and the hypctx structure is used to save guest
registers.  The hyp structure mapping is currently not used to store
capabilities, but set VM_PROT_WRITE_CAP there anyway to avoid nasty
surprises in the future.