libgeom: Avoid fixed remappings of the devstat device
libgeom maintains a quasi-private mapping of /dev/devstat, which might grow
over time if new devices appear.  When the mapping needs to be expanded, it
maps over the old region with MAP_FIXED, but there is no readily apparent
reason for doing it this way.

mmap(2) fails with EPROT when used this way under CheriABI, and more
generally it seems wrong since growing the mapping this way may trample over
something else.

Thus:
- stop using MAP_FIXED,
- don't creat a mapping in geom_stats_open(), as geom_stats_resync() will
  create it for us,
- check for errors from munmap().