linker: Improve handling of ifunc symbols
When linker_file_function_listall() enumerates function symbols on purecap
platforms, ifunc symbol values are sealed, whereas regular function symbols are
not.  This breaks dtrace's fbt and kinst providers, which use the returned
capability to disassemble each function in the kernel.  Moreover, the size of
the symbol is wrong: it's the size of the resolver function rather than that of
the implementation function.  This latter problem is a bug upstream too.

Fix both problems by looking up the full ELF symbol for the resolved function.

This change is not quite correct, as it assumes that ifunc resolvers always
return a symbol within the same linker file.  However, this is true in practice
today.