vm_fault: Fix handling of VM_FAULT_NOPMAP
This flag, which is CheriBSD-specific, indicates that vm_fault() is not to
create a mapping of the faulted page.  There are several places where a mapping
can be created, and VM_FAULT_NOPMAP is applied in all but one.

Be consistent and apply it here too, as there's no clear reason not to, and the
semantics of VM_FAULT_NOPMAP are otherwise not clear.