vm_fault: Allow some prefaulting during revocation
The loop in vm_fault_prefault() which actually maps pages will also check
whether revocation is required before creating the mapping.  That check is more
nuanced and doesn't exclude some cases where it's fine to create the mapping
even if revocation is ongoing, for example when pre-faulting a mapping of a
vnode.